change project architectur and satrt aave implementation
