+++ back-end/app/swagger.json
@@ -78,4 +78,4 @@
             }
         }
     }
-}
+}
