+++ back-end/package.json
@@ -2,8 +2,9 @@
   "name": "back-end",
   "version": "1.0.0",
   "description": "",
-  "main": "server.js",
+  "main": "./app/server.js",
   "scripts": {
+    "start": "node ./app/server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
